basic structure for the yara service.
