+++ src/main/scala/org/novetta/zoo/driver/driver.scala
@@ -4,8 +4,8 @@ import java.util.concurrent.{Executors, ExecutorService}
 
 import akka.actor.{ActorRef, ActorSystem, Props}
 import org.novetta.zoo.actors._
-import org.novetta.zoo.services.peinfo.{PEInfoSuccess, PEInfoWork}
-import org.novetta.zoo.services.{YaraSuccess, MetadataSuccess, VTSampleSuccess, YaraWork, MetadataWork, VTSampleWork}
+import org.novetta.zoo.services.yara.{YaraSuccess, YaraWork}
+import org.novetta.zoo.services.{MetadataSuccess, VTSampleSuccess, MetadataWork, VTSampleWork}
 import org.novetta.zoo.types._
 
 import org.json4s._
