+++ src/main/scala/org/novetta/zoo/services/yara/Dockerfile
@@ -2,12 +2,12 @@ FROM ubuntu:14.04
 
 # install what you need here
 RUN apt-get update
-RUN apt-get install -y python-pip
-RUN pip install tornado pefile bitstring
+RUN apt-get install -y --fix-missing yara
+RUN pip install yara-python
 
 # add the files to the container
 RUN mkdir -p /service
-ADD peinfo.py /service
+ADD yara.py /service
 
 # create a new user with limited access
 RUN useradd -s /bin/bash service
@@ -16,5 +16,5 @@ RUN chown -R service /service
 USER service
 WORKDIR /service
 
-EXPOSE 7705
-CMD ["python2", "/service/peinfo.py"]
+EXPOSE 7701
+CMD ["python2", "/service/yara.py"]
