+++ src/main/scala/org/novetta/zoo/services/yara/README.md
