+++ src/main/scala/org/novetta/zoo/services/yara/acl.conf
