+++ src/main/scala/org/novetta/zoo/services/yara/service.conf
