+++ src/main/scala/org/novetta/zoo/services/yara/watchguard.scala
